# typescript/packages/apps/reference-datasets/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIF SaaS: Reference Datasets Management",
    "description": "\nHas accountability for:\n- TODO...\n",
    "version": "0.0.1"
  },
  "components": {
    "securitySchemes": {
      "platformUserPool": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "signedUrl_response": {
        "type": "object",
        "properties": {
          "url": {
            "description": "requested signed url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ]
      }
    }
  },
  "paths": {
...
    "/referenceDatasets/{id}/versions/{version}/data": {
      "get": {
        "operationId": "getVersionData",
        "tags": [
          "Reference Datasets"
        ],
        "description": "Retrieve specific version of reference dataset file content\n\nPermissions:\n- All roles of the group in context may get reference dataset.\n",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Unique id."
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "version",
            "required": true,
            "description": "version of the reference dataset"
          },
          {
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type",
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-security-scopes": [
          "reader",
          "contributor",
          "admin",
          "superAdmin"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "text/csv": {
                "schema": {
                  "description": "Success.",
                  "$ref": "#/components/schemas/signedUrl_response"
                },
                "examples": {
                  "Content of the reference dataset": {
                    "summary": "Content of the reference dataset",
                    "value": "Header1,Header2\nRow1Column1,Row1Column2\nRow2Column1,Row2Column2"
                  },
                  "Signed Url for the Reference dataset": {
                    "summary": "Signed url to retrieve the reference dataset",
                    "value": {
                      "url": "https://signed.url.response/data.object"
                    }
                  }
                }
              },
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "$ref": "#/components/schemas/signedUrl_response"
                },
                "examples": {
                  "Content of the reference dataset": {
                    "summary": "Content of the reference dataset",
                    "value": "Header1,Header2\nRow1Column1,Row1Column2\nRow2Column1,Row2Column2"
                  },
                  "Signed Url for the Reference dataset": {
                    "summary": "Signed url to retrieve the reference dataset",
                    "value": {
                      "url": "https://signed.url.response/data.object"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "text/csv": {
                "schema": {
                  "description": "Forbidden.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "description": "Forbidden.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "text/csv": {
                "schema": {
                  "description": "Not found.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "description": "Not found.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/referenceDatasets/{id}/versions/{version}/index": {
      "get": {
        "operationId": "getVersionIndex",
        "tags": [
          "Reference Datasets"
        ],
        "description": "Retrieve specific version of reference dataset lucene index\n\nPermissions:\n- All roles of the group in context may get reference dataset.\n",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Unique id."
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "version",
            "required": true,
            "description": "version of the reference dataset"
          },
          {
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type",
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-security-scopes": [
          "reader",
          "contributor",
          "admin",
          "superAdmin"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "$ref": "#/components/schemas/signedUrl_response"
                },
                "examples": {
                  "Signed Url for the Reference dataset index file": {
                    "summary": "Signed url to retrieve the reference dataset index file",
                    "value": {
                      "url": "https://signed.url.response/data.object"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Forbidden.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Not found.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "security": [],
...
}